refactor(trie): break down `trie/node` into smaller files and renaming
