+++ packages/trie/src/util/nibbles.ts
@@ -1,4 +1,4 @@
-import { Nibbles } from '../trie'
+import { Nibbles } from '../types'
 
 /**
  * Converts a buffer to a nibble array.
