+++ packages/trie/src/util/walkController.ts
@@ -1,7 +1,7 @@
 import { Trie } from '../trie'
-import { FoundNodeFunction } from '../types'
+import { FoundNodeFunction, Nibbles, TrieNode } from '../types'
 import { PrioritizedTaskExecutor } from './tasks'
-import { BranchNode, ExtensionNode, LeafNode, Nibbles, TrieNode } from '../trie/node'
+import { BranchNode, ExtensionNode, LeafNode } from '../trie/node'
 
 /**
  * WalkController is an interface to control how the trie is being traversed.
