+++ packages/trie/tsconfig.json
@@ -3,6 +3,5 @@
   "compilerOptions": {
       "outDir": "./dist",
     },
-  "include": ["src/**/*.ts", "test/*.spec.ts"]
+  "include": ["src/**/*.ts", "test/**/*.spec.ts"]
 }
-
